+++ packages/frontend/src/components/Settings/Profile.tsx
@@ -7,12 +7,15 @@ import LargeProfileImage from '../LargeProfileImage'
 import type { SettingsStoreState } from '../../stores/settings'
 
 import styles from './styles.module.scss'
+import useTranslationFunction from '../../hooks/useTranslationFunction'
 
 type Props = {
   settingsStore: SettingsStoreState
 }
 
 export default function Profile({ settingsStore }: Props) {
+  const tx = useTranslationFunction()
+
   const initials = avatarInitial(
     settingsStore.settings?.displayname || '',
     settingsStore.selfContact.address
@@ -33,8 +36,9 @@ export default function Profile({ settingsStore }: Props) {
         <div className={styles.profileDisplayName}>
           {settingsStore.settings.displayname}
         </div>
-        <div className={styles.profileAddress}>
-          {settingsStore.selfContact.address}
+        <div className={styles.profileBio}>
+          {settingsStore.settings.selfstatus?.replace('\n', ' ') ||
+            tx('pref_default_status_label')}
         </div>
       </div>
     </div>