+++ packages/frontend/src/components/Settings/SettingsIconButton.tsx
@@ -11,6 +11,7 @@ type Props = PropsWithChildren<{
   icon: IconName
   isLink?: boolean
   onClick: () => void
+  dataTestid?: string
 }>
 
 export default function SettingsIconButton({
@@ -18,9 +19,14 @@ export default function SettingsIconButton({
   icon,
   isLink = false,
   onClick,
+  dataTestid,
 }: Props) {
   return (
-    <button className={styles.settingsIconButton} onClick={onClick}>
+    <button
+      className={styles.settingsIconButton}
+      onClick={onClick}
+      data-testid={dataTestid}
+    >
       <Icon className={styles.settingsIcon} icon={icon} />
       <span className={styles.settingsIconButtonLabel}>{children}</span>
       {isLink && <Icon className={styles.settingsIcon} icon='open_in_new' />}