+++ packages/target-tauri/src-tauri/src/state/main_window_channels.rs
@@ -33,6 +33,7 @@ pub enum MainWindowEvents {
     ShowKeybindingsDialog,
     ResumeFromSleep,
     ToggleNotifications,
+    OnThemeUpdate,
 }
 
 pub(crate) struct InnerMainWindowChannelsState {